pyproject.toml
[project]
name = "telegram-knowledge-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.116.1",
    "httpx>=0.28.1",
    "langchain-community>=0.3.27",
    "pydantic>=2.11.5",
    "pydantic-ai-slim[duckduckgo]>=0.4.2",
    "pydantic-graph>=0.3.1",
    "pydantic-settings>=2.9.1",
    "stravalib>=2.1.2",
    "python-dotenv>=1.0.1",
    "alembic>=1.16.4",
    "sqlalchemy>=2.0.41",
    "psycopg[binary]>=3.2.9",
    "langchain-google-community[calendar]>=2.0.7",
    "mem0ai>=0.1.116",
]
